+++ base/src/main/java/io/github/sds100/keymapper/base/constraints/ConstraintSnapshot.kt
@@ -57,6 +57,9 @@ class LazyConstraintSnapshot(
         networkAdapter.connectedWifiSSIDFlow.firstBlocking()
     }
     private val chosenImeId: String? by lazy { inputMethodAdapter.chosenIme.value?.id }
+    private val isKeyboardShowing: Boolean by lazy {
+        accessibilityService.isInputMethodVisible.firstBlocking()
+    }
     private val callState: CallState by lazy { phoneAdapter.getCallState() }
     private val isCharging: Boolean by lazy { powerAdapter.isCharging.value }
 
@@ -139,6 +142,8 @@ class LazyConstraintSnapshot(
             is ConstraintData.WifiOn -> isWifiEnabled
             is ConstraintData.ImeChosen -> chosenImeId == constraint.data.imeId
             is ConstraintData.ImeNotChosen -> chosenImeId != constraint.data.imeId
+            is ConstraintData.KeyboardShowing -> isKeyboardShowing
+            is ConstraintData.KeyboardNotShowing -> !isKeyboardShowing
             is ConstraintData.DeviceIsLocked -> isLocked
             is ConstraintData.DeviceIsUnlocked -> !isLocked
             is ConstraintData.InPhoneCall ->