+++ base/src/test/java/io/github/sds100/keymapper/base/actions/PerformActionsUseCaseTest.kt
@@ -86,6 +86,8 @@ class PerformActionsUseCaseTest {
             systemBridgeConnectionManager = mock(),
             executeShellCommandUseCase = mock(),
             coroutineScope = testCoroutineScope,
+            notificationAdapter = mock(),
+            settingsAdapter = mock(),
         )
     }